chore: InputForm and SearchInput components merged to Input component
