+++ src/assets/Icons/icons/SearchIcon.tsx
@@ -7,7 +7,7 @@ export const SearchIcon = ({size}:IconProps) => {
   const theme = useTheme();
 
   return (
-    <svg width={size} height={size} viewBox="0 0 24 24" fill="none" xmlns="http://www.w3.org/2000/svg">
+    <svg id='search' width={size} height={size} viewBox="0 0 24 24" fill="none" xmlns="http://www.w3.org/2000/svg">
       <g id="Interface / Search_Magnifying_Glass">
         <path
           id="Vector"
