+++ src/index.ts
@@ -1,2 +1,2 @@
 export * from './components/Button';
-export * from './components/InputForm';
+export * from './components/Input';
